pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "plugins"]

[project]
name = "granite"
version = "0.1.0"
description = "A modern note-taking application built with FastAPI"
requires-python = ">=3.10"
dependencies = [
    # Core FastAPI dependencies
    "fastapi>=0.115.0",
    "starlette>=0.47.2",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "filelock>=3.20.3",
    # Content processing
    "markdown==3.7",
    "pyyaml==6.0.1",
    # Async file operations
    "aiofiles==24.1.0",
    # Security
    "cryptography==44.0.1",
    "bcrypt==4.2.1",
    "itsdangerous==2.2.0",
    # Data validation
    "pydantic>=2.0.0",
    # Logging
    "loguru==0.7.2",
    # Rate limiting
    "slowapi==0.1.9",
    # Testing
    "pytest==8.3.3",
    "httpx==0.28.0",
    "pytest-cov==6.0.0",
    # PDF Export Plugin dependencies
    "weasyprint==63.1",
    "pydyf==0.11.0",
    "Pygments==2.18.0",
    # Code quality tools (development)
    "ruff==0.14.9",
    "pre-commit>=4.5.1",
    "yamllint>=1.38.0",
    "bandit>=1.9.2",
    "mypy>=1.19.1",
]

# Ruff configuration - replaces black, flake8, isort, and more
[tool.ruff]
line-length = 120
target-version = "py310"

# Exclude directories
extend-exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "data",
]

[tool.ruff.lint]
# Enable these rule sets
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "PL",     # pylint
    "TRY",    # tryceratops
    "RUF",    # ruff-specific rules
]

# Ignore specific rules
ignore = [
    "E501",    # Line too long (handled by formatter)
    "PLR0913", # Too many arguments
    "PLR0912", # Too many branches
    "PLR0915", # Too many statements
    "PLR2004", # Magic value in comparison (unavoidable for HTTP status codes, etc.)
    "TRY003",  # Avoid specifying long messages outside the exception class
    "TRY300",  # Consider moving statement to else block (overly opinionated)
    "TRY301",  # Abstract raise to inner function (overly opinionated)
    "EM101",   # Exception must not use a string literal
    "EM102",   # Exception must not use an f-string literal
    "ARG001",  # Unused function argument (common in FastAPI routes)
    "ARG002",  # Unused method argument (common in base classes/plugins)
    "B008",    # Function call in argument defaults (required for FastAPI dependency injection)
    "PT028",   # Test function parameter defaults (false positive for FastAPI patterns)
    "PLC0415", # Import not at top-level (lazy imports for performance/debugging are intentional)
]

# Allow fix for all enabled rules (when `--fix` is provided)
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
# Ignore certain rules in test files
"tests/**/*.py" = [
    "ARG",     # Unused arguments OK in tests
    "PT",      # pytest-style rules
]

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Indent with spaces
indent-style = "space"

# Like Black, respect magic trailing commas
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending
line-ending = "auto"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--strict-config",
    "--cov=backend",
    "--cov=plugins",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["backend", "plugins"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]  # B101: assert_used, B601: paramiko_calls

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter typing
ignore_missing_imports = true
